isolated FormPanes from App.js
